% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_book.R
\name{build_current_book_part_docx}
\alias{build_current_book_part_docx}
\title{Generating a 'bookdown::word_document2' document from the current
edited bookdown chapter (Rmd)}
\usage{
build_current_book_part_docx()
}
\description{
Generating an output document from the open Rmd file
in the RStudio editor without specifying neither the file name nor
the output format ('bookdown::word_document2')
}
\examples{
\dontrun{

build_current_book_part_docx()

}


}
\seealso{
\code{\link{generate_book}} for specifying Rmd files list, and output
format
}
