% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_book.R
\name{build_current_book_part_pdf}
\alias{build_current_book_part_pdf}
\title{Generating a 'bookdown::gitbook' document from the current
edited bookdown chapter (Rmd)}
\usage{
build_current_book_part_pdf()
}
\description{
Generating an output document from the open Rmd file
in the RStudio editor without specifying neither the file name nor
the output format ('bookdown::gitbook')
}
\examples{
\dontrun{

build_current_book_part_html()

}

}
\seealso{
\code{\link{generate_book}} for specifying Rmd files list, and output
format
}
