% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stics_names.R
\name{find_stics_names}
\alias{find_stics_names}
\title{Gadget for finding Stics names (parameters, variables)}
\usage{
find_stics_names()
}
\description{
Gadget for finding Stics names (parameters, variables)
}
