% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{gen_bibtex_file}
\alias{gen_bibtex_file}
\title{Generating a BibTex file using a query to a Zotero web library}
\usage{
gen_bibtex_file(
  file_dir = getwd(),
  library_name,
  collection_name = NULL,
  query = "",
  query_type = "full",
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{file_dir}{Optional, a BibTex file directory path (default: current)}

\item{library_name}{A group library name (appearing in groups in Zotero web)}

\item{collection_name}{Optional, a collection name (or vector of) belonging to the library}

\item{query}{Optional, a character string to use for querying bibliographic base}

\item{query_type}{Optional, a query type: "full" (the default) for searching
in all the fields, or "basic" for searching only in title, creator, year fields}

\item{verbose}{Optional, a logical used for displaying messages
from the query process (TRUE) or not (FALSE)}

\item{overwrite}{Optional, a logical for deleting existing file (TRUE),
or not (FALSE, default)}
}
\value{
An invisible BibEntry object or a list of
}
\description{
Generating a BibTex file using a query to a Zotero web library
}
\details{
The output file name(s) is/are either the library name (if not any collection are
given) or the collection name(s)
}
\examples{
\dontrun{

# No collection
gen_bibtex_file(library_name = "GroupLibraryName")

# With a collection name
gen_bibtex_file(library_name = "GroupLibraryName", collection_name = "CollName")

# With query, searching in all fields
gen_bibtex_file(library_name = "GroupLibraryName",
collection_name = "CollName", query = "1995" )

# Getting returned BibEntry object
bibentry_obj <- gen_bibtex_file(library_name = "GroupLibraryName")

}


}
