% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{generate_bibfile}
\alias{generate_bibfile}
\title{Generating a BibTex file using a query to a Zotero web base}
\usage{
generate_bibfile(
  file = NULL,
  query = "",
  query_mode = "full",
  start = 1,
  verbose = FALSE
)
}
\arguments{
\item{file}{A BibTex file path (with a .bib extension)}

\item{query}{A character string to use for querying bibliographic base}

\item{query_mode}{A query mode: "full" (the default) for searching in all the fields,
or "basic" for searching only in title, creator, year fields}

\item{start}{A numeric indicating from which record number in the base to start the query}

\item{verbose}{A logical for displayin messages from the query process (TRUE) or not (FALSE)}
}
\description{
Generating a BibTex file using a query to a Zotero web base
}
\examples{
\dontrun{
generate_bibfile("file.bib")

generate_bibfile(file = "file.bib", query = "1995" )

generate_bibfile(file = "file.bib", query = "1995", query_mode = "basic" )

}


}
