% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_tools.R
\name{plot_render}
\alias{plot_render}
\title{Saving a plot (i.e. ggplot object) to an image file}
\usage{
plot_render(
  plot_obj,
  img_name = NULL,
  img_format = "png",
  img_dir = getwd(),
  img_save = FALSE,
  opts = NULL
)
}
\arguments{
\item{plot_obj}{A plot object or function}

\item{img_name}{File name}

\item{img_format}{Image file format}

\item{img_dir}{Output file directory}

\item{img_save}{Logical for saving a plot in a file or not (default)}

\item{opts}{a knitr::opts_chunk object containing all above img_* arguments values
as elements}
}
\value{
a plot object
}
\description{
Saving a plot (i.e. ggplot object) to an image file
}
\examples{
\dontrun{
 plot_render(plot_obj)
 plot_render(plot_obj, img_name = "myimage", img_save = TRUE)

}


}
