% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{update_stics_bibtex_files}
\alias{update_stics_bibtex_files}
\title{Generating or updating silently a BibTex file or a list of,
using a Zotero web library}
\usage{
update_stics_bibtex_files(
  file_dir = getwd(),
  library_name = "livre_rouge",
  collection_name = c("sticsredbook2020", "EPS_stics"),
  query = "",
  query_type = "full"
)
}
\arguments{
\item{file_dir}{Optional, a BibTex file directory path (default: current)}

\item{library_name}{A group library name (appearing in groups in Zotero web)}

\item{collection_name}{Optional, a collection name (or vector of) belonging to the library}

\item{query}{Optional, a character string to use for querying bibliographic base}

\item{query_type}{Optional, a query type: "full" (the default) for searching
in all the fields, or "basic" for searching only in title, creator, year fields}
}
\value{
An invisible BibEntry object or a list of
}
\description{
Generating or updating silently a BibTex file or a list of,
using a Zotero web library
}
\examples{
\dontrun{

# For the default collection names vector
update_stics_bibtex_files()

# For the hole library_name
update_stics_bibtex_files(collection_name = NULL)

}


}
