% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{gen_bibtex_file}
\alias{gen_bibtex_file}
\title{Generating a BibTex file using a query to a Zotero web base}
\usage{
gen_bibtex_file(
  file_path = NULL,
  query = "",
  query_type = "full",
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{file_path}{A BibTex file path (with a .bib extension)}

\item{query}{A character string to use for querying bibliographic base}

\item{query_type}{A query type: "full" (the default) for searching in all the fields,
or "basic" for searching only in title, creator, year fields}

\item{verbose}{A logical for displaying messages from the query process (TRUE) or not (FALSE)}

\item{overwrite}{A logical for deleting existing file (TRUE), or not (FALSE, default)}
}
\description{
Generating a BibTex file using a query to a Zotero web base
}
\examples{
\dontrun{

# No query
gen_bibtex_file("file.bib")

# with query, searching in all fields
gen_bibtex_file(file = "file.bib", query = "1995" )

# searching in basic fields (title, creator, year)
gen_bibtex_file(file = "file.bib", query = "1995", query_type = "basic" )

# getting returned BibEntry object
bibentry_obj <- gen_bibtex_file("file.bib")

}


}
