% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_book.R
\name{generate_book}
\alias{generate_book}
\title{Generating a bookdown document or a part of}
\usage{
generate_book(
  input_rmd = NULL,
  book_pkg_dir = getwd(),
  output_format = NULL,
  params = list()
)
}
\arguments{
\item{input_rmd}{Chapter Rmd file name (optional)}

\item{book_pkg_dir}{Book package directory path (optional, default: current directory)}

\item{output_format}{String defining the type of document to build
(optional, default: "bookdown::html_document2")}

\item{params}{A list containing parameters to pass to the Rmd
document to parameterize the output or overload existing params in the
document}
}
\value{
A success status of the document generation
}
\description{
The generation of an entire book is performed without
any given arguments using the current directory which must be an
package book directory or giving a path to the package. A part of the
book may be generated when a Rmd file name or a list of is given.
}
\examples{
\dontrun{
generate_book()

generate_book(input_rmd = "07-chap.Rmd")

generate_book(input_rmd = "07-chap.Rmd", output_format = 'bookdown::pdf_document2')

}



}
