% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_book.R
\name{generate_book}
\alias{generate_book}
\title{Generating a bookdown document or a part of}
\usage{
generate_book(chap_rmd = NULL, book_pkg_dir = getwd())
}
\arguments{
\item{chap_rmd}{Chapter Rmd file name (optional)}

\item{book_pkg_dir}{}
}
\value{
A status of the generated document
}
\description{
The generation of an entire book is performed without
any given arguments using the current directory which must be an
package book directory or giving a path to the package. A part of the
book may be generated when a Rmd file name is given.
}
